+++ userspace/lib/include/sys/mman.h
@@ -23,9 +23,24 @@
 #define MS_SYNC SRV_MS_SYNC
 #define MS_INVALIDATE SRV_MS_INVALIDATE
 
+#define MADV_NORMAL 0
+#define MADV_RANDOM 1
+#define MADV_SEQUENTIAL 2
+#define MADV_WILLNEED 3
+#define MADV_DONTNEED 4
+#define MADV_FREE 8
+#define MADV_REMOVE 9
+#define MADV_DONTFORK 10
+#define MADV_DOFORK 11
+#define MADV_MERGEABLE 12
+#define MADV_UNMERGEABLE 13
+#define MADV_HUGEPAGE 14
+#define MADV_NOHUGEPAGE 15
+
 void *mmap(void *address, size_t length, int protection, int flags, int fd, off_t offset);
 int munmap(void *address, size_t length);
 int mprotect(void *address, size_t length, int protection);
 int msync(void *address, size_t length, int flags);
+int madvise(void *address, size_t length, int advice);
 
 #endif