+++ userspace/lib/include/sys/sysinfo.h
@@ -1,0 +1,23 @@
+#ifndef SRVROS_POSIX_SYS_SYSINFO_H
+#define SRVROS_POSIX_SYS_SYSINFO_H
+
+#include <stdint.h>
+
+struct sysinfo {
+    long uptime;
+    unsigned long loads[3];
+    unsigned long totalram;
+    unsigned long freeram;
+    unsigned long sharedram;
+    unsigned long bufferram;
+    unsigned long totalswap;
+    unsigned long freeswap;
+    unsigned short procs;
+    unsigned long totalhigh;
+    unsigned long freehigh;
+    unsigned int mem_unit;
+};
+
+int sysinfo(struct sysinfo *info);
+
+#endif