+++ userspace/lib/src/posix_linux_compat.c
@@ -1,0 +1,150 @@
+#include <errno.h>
+#include <stdarg.h>
+#include <stdint.h>
+#include <stdio.h>
+#include <string.h>
+#include <sys/auxv.h>
+#include <sys/mman.h>
+#include <sys/prctl.h>
+#include <sys/random.h>
+#include <sys/syscall.h>
+#include <sys/sysinfo.h>
+#include <unistd.h>
+
+#include <srvros/sys.h>
+
+int sysinfo(struct sysinfo *info) {
+    if (info == 0) {
+        errno = EINVAL;
+        return -1;
+    }
+    memset(info, 0, sizeof(*info));
+    struct srv_meminfo meminfo;
+    if (srv_meminfo(&meminfo) < 0) {
+        errno = EIO;
+        return -1;
+    }
+    long ticks_per_second = sysconf(_SC_CLK_TCK);
+    if (ticks_per_second <= 0) {
+        ticks_per_second = 100;
+    }
+    info->uptime = (long)(srv_ticks() / (uint64_t)ticks_per_second);
+    info->totalram = (unsigned long)meminfo.total_bytes;
+    info->freeram = (unsigned long)meminfo.free_bytes;
+    info->mem_unit = 1;
+    info->procs = 1;
+    return 0;
+}
+
+unsigned long getauxval(unsigned long type) {
+    switch (type) {
+    case AT_PAGESZ:
+        return (unsigned long)getpagesize();
+    case AT_CLKTCK:
+        return (unsigned long)sysconf(_SC_CLK_TCK);
+    case AT_SECURE:
+    case AT_HWCAP:
+    case AT_HWCAP2:
+    case AT_SYSINFO_EHDR:
+        return 0;
+    default:
+        errno = ENOENT;
+        return 0;
+    }
+}
+
+int prctl(int option, ...) {
+    switch (option) {
+    case PR_SET_NAME:
+    case PR_SET_VMA:
+        return 0;
+    case PR_GET_NAME: {
+        va_list args;
+        va_start(args, option);
+        char *buffer = va_arg(args, char *);
+        va_end(args);
+        if (buffer == 0) {
+            errno = EINVAL;
+            return -1;
+        }
+        strcpy(buffer, "srvros");
+        return 0;
+    }
+    default:
+        errno = EINVAL;
+        return -1;
+    }
+}
+
+long syscall(long number, ...) {
+    va_list args;
+    va_start(args, number);
+    long result = -1;
+    switch (number) {
+    case SYS_read: {
+        int fd = va_arg(args, int);
+        void *buffer = va_arg(args, void *);
+        size_t length = va_arg(args, size_t);
+        result = read(fd, buffer, length);
+        break;
+    }
+    case SYS_write: {
+        int fd = va_arg(args, int);
+        const void *buffer = va_arg(args, const void *);
+        size_t length = va_arg(args, size_t);
+        result = write(fd, buffer, length);
+        break;
+    }
+    case SYS_close: {
+        int fd = va_arg(args, int);
+        result = close(fd);
+        break;
+    }
+    case SYS_mmap: {
+        void *address = va_arg(args, void *);
+        size_t length = va_arg(args, size_t);
+        int protection = va_arg(args, int);
+        int flags = va_arg(args, int);
+        int fd = va_arg(args, int);
+        off_t offset = va_arg(args, off_t);
+        void *mapping = mmap(address, length, protection, flags, fd, offset);
+        result = mapping == MAP_FAILED ? -1 : (long)(uintptr_t)mapping;
+        break;
+    }
+    case SYS_mprotect: {
+        void *address = va_arg(args, void *);
+        size_t length = va_arg(args, size_t);
+        int protection = va_arg(args, int);
+        result = mprotect(address, length, protection);
+        break;
+    }
+    case SYS_munmap: {
+        void *address = va_arg(args, void *);
+        size_t length = va_arg(args, size_t);
+        result = munmap(address, length);
+        break;
+    }
+    case SYS_getpid:
+        result = getpid();
+        break;
+    case SYS_gettid:
+        result = srv_thread_self();
+        if (result <= 0) {
+            result = getpid();
+        }
+        break;
+    case SYS_getrandom: {
+        void *buffer = va_arg(args, void *);
+        size_t length = va_arg(args, size_t);
+        unsigned int flags = va_arg(args, unsigned int);
+        result = getrandom(buffer, length, flags);
+        break;
+    }
+    default:
+        errno = ENOSYS;
+        result = -1;
+        break;
+    }
+    va_end(args);
+    return result;
+}